Task 1: Solved The First Task
Implemented separate functions for reading, printing and freeing the dynamically alloted
space for the generation matrix. Added two main functions, one that changes the generation
matrix according to the Game of Life rules, using the other one that calculates the number
of alive neighbours. The number of neighbours is then memorised in a matrix parallel to the
generation matrix, so that the changes don't happen before the whole generation is checked.

Signed-off-by: Homm Filip-Razvan <filip_razvan.homm@stud.upb.acs.ro>